% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_moments.R
\name{align_moments}
\alias{align_moments}
\title{Align NEXRAD Moments}
\usage{
align_moments(nc_file, csv_out = NULL)
}
\arguments{
\item{nc_file}{NetCDF File}

\item{csv_out}{File path to save output}
}
\value{

}
\description{
Align NEXRAD Moments
}
