% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nexrad.R
\name{download_nexrad}
\alias{download_nexrad}
\title{Download NEXRAD}
\usage{
download_nexrad(station_id, start_datetime, end_datetime = NULL)
}
\arguments{
\item{station_id}{The 4 letter NEXRAD station code}

\item{start_datetime}{A character vector representing the UTC start time of the data request. Format: '%Y-%m-%d %H:%M:%S'}

\item{end_datetime}{A character vector representing the UTC end time of the data request. Format: '%Y-%m-%d %H:%M:%S'}
}
\value{
A list of downloaded NEXRAD files
}
\description{
Download NEXRAD
}
